requirements.txt
dataclasses==0.1
Django==2.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
djongo==1.2.30
numpy==1.18.4
onnx==1.7.0
onnxruntime==1.2.0
opencv-python==4.2.0.34
protobuf==3.12.0
pymongo==3.7.1
pytz==2018.5
six==1.14.0
sqlparse==0.2.3
typing-extensions==3.7.4.2
